lib: internal: worker: packages: __generate_input_for_preinstall_v2_hooks: do not skip sign when compareVersionStrings is not 1, -1, or 0
In some cases compareVersionStrings passes on the return value from
memcmp, which is not guaranteed to be 1, -1, or 0. Most callers don't
mind, but one does: the value gets used in a switch statement to set
the comparison symbol for use in VERSION 2 apt pre-installation hook
output, and since no case label matches, the comparison symbol string
is left empty, resulting in hook output with lines like

	libgnutls26 2.12.5-1  2.12.7-1 **CONFIGURE**

where there should be

	libgnutls26 2.12.5-1 < 2.12.7-1 **CONFIGURE**

and apt-listbugs crashing with the message

	E: private method `chomp!' called for nil:NilClass

because it sees 4 fields where there should be 5. Fix it by only
checking the sign of the return value from compareVersionStrings
instead of switching on its value.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Eugene V. Lyubimkin <jackyf.devel@gmail.com>