lib: system: worker: new method 'setPackagePurgeFlag'
From now on, front-end can specify which removed packages are to be
purged and which are not, one by one.

The global purge option 'cupt::worker::purge' is restored to its
original behavior to purge all packages marked for removal. On the other
hand, this option is no more tied to front-end's 'purge'/'--purge'
action/option.