lib: internal: worker: packages: __build_actions_graph: do not refill dependencies on each iteration
The commit '6f5bec10' makes it possible to delete edges which are not
needed anymore due to fully linked chains, since the candidate edges are
not deleted anymore but saved with the needed relation information after
the first run of __fill_action_dependencies().

This commit moves checking for edges that can be deleted to
__expand_linked_actions(), which is still called iteratively.

What we gain is graph doesn't have to be refilled anymore on each
iteration.