lib: internal: worker: packages: __set_priority_links: don't adjust when the chain is fully linked already
The only purpose of this function is to help chains be
"walkable-through", i.e., dependencies for non-first subactions are
ready before the first subaction is started, so the subactions may be
linked.

When the edge 'to' -> 'from' is present it means that all subactions are
linked already because they belong to the same subgroup. In that case
there is no need to set anything.