lib: internal: worker: packages: __expand_linked_actions, __build_mini_action_graph: transfer/restore potential edges
Actions are linked for the goal that some two actions, each of one of
two different packages, are not edged anymore if it's possible that the
"dependent" package is fully chained.

However it only works only when no actions of these different packages
is mixed in the next-level subactions ("no-mix-in-next"). Failing that,
hard dependencies in the rare cases may be overridden. This is a
long-standing bug (since 2.0.0~alpha3), but it didn't show up until the
commit '1796911f' where splitting actions in several phases instead of
one was introduced.

To overcome that, edges that are deleted in the process of linking, are
saved to the location 'potentialEdges'. When mini action graphs are
building we check edges to be deleted and unlink (i.e. restore) the
edges which fail to satisfy that 'no-mix-in-next' rule.