lib: internal: worker: base: do not redirect commands' input unnecessarily
Commit 'fbdc779d' taught the worker to submit input for hooks and
other commands that require special data on stdin through a pipe
instead of asking the shell to redirect for us.  To keep the code
simple, it taught cupt to _always_ make stdin of external commands a
pipe, with the unpleasant consequence that the user cannot interact
via the terminal even if there was no input to be passed from cupt:

	| # cupt upgrade highlight-common
	| [...]
	| Unpacking highlight-common (from .../highlight-common_2.16-1_all.deb) ...
	| Setting up highlight-common (2.16-1) ...
	|
	| Configuration file `/etc/highlight/filetypes.conf'
[...]
	|  The default action is to keep your current version.
	| *** filetypes.conf (Y/I/N/O/D/Z) [default=N] ? dpkg: error processing highlight-common (--install):
	|  EOF on stdin at conffile prompt
	| Errors were encountered while processing:
	|  highlight-common
	| E: pipe '/usr/bin/dpkg --install --no-triggers --force-bad-path //var/cache/apt/archives/highlight-common_2.16-1_all.deb' execution failed: exit code '1'

"Fix" it by only redirecting stdin when the commandInput parameter is
nonempty.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Eugene V. Lyubimkin <jackyf.devel@gmail.com>