lib: do not suppress default copy constructors and assignment operators
Building cupt with g++ 4.6 produces some errors due to missing
copy-assignment operators and copy constructors.

The C++ standard (n3242, section special.class.copy "Copying and
moving class objects", paragraph 8) explains:

	If the class definition does not explicitly declare a copy
	constructor, there is no user-declared move constructor, and
	there is no user-declared move assignment operator, a copy
	constructor is implicitly declared as defaulted.

In other words, if we declare _any_ move constructor or move
assignment operator, the default copy constructor is not guaranteed to
be present.  (Likewise for the default copy-assignment operator.)

Worse, most of the move operations we provide are superfluous, since
they just call std::move to convert "operand.member" from an lvalue to
an xvalue, and members of an object referred to by rvalue reference
are already xvalues.  Simplify by using the implicit definitions that
accomplish the same thing (but to be cautious, explicitly declare that
we are doing so).

The only exception is cupt::internal::PackageEntry, whose rvalue
operations explicitly call std::vector::swap.  That is probably not
needed, either, since std::vector provides rvalue operations, too, but
to be conservative, this patch leaves it alone.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>

Conflicts:

	cpp/lib/src/internal/nativeresolver/solution.hpp

Signed-off-by: Eugene V. Lyubimkin <jackyf.devel@gmail.com>