lib: internal: nativeresolver: impl: __clean_automatically_installed: fully revert auto-discarding solutions
Auto-discarding solutions turned to cause sometimes far more serious
problems than it solves.

Particularly, in the middle of complex situation, if one of the
solutions to the problem is the removal of automatically installed
package and the solution decided to not remove the element (usually,
because a solution with a bigger is available), all successor solutions
will fail. Similarly for partial sticking.

I believe there's no way to tell in the middle of a process
will the package be auto-removed or not. Even if we decide to always
favor auto-removal over other solutions, there is always a possibility
that some new package (which will be installed later) will depend on
auto-removed one.

This commit again opens a possibility for native resolver to return
"duplicated" solutions, but the solutions are not (and were not in the
first place) full duplicates if we count also the reasons of changes.
As the console front-end is now able to distinct between autoremovals
and resolver-caused removals, I believe this change should not raise the
'why the duplicates?' question again.