lib: internal: nativeresolver: solution: SolutionStorage: setPackageEntry: fixed spurious 'conflicting elements' internal error
Fixes the regression which was introduced in the commit
'73c89d0be832cf3be10633dd3acca59c07950902'.

After that commit it's possible that a deleted element is physically
present in the __added_entries.

The fix required two changes:

1) The internal error precondition had to be relaxed, because it's legal
for the autoremover to reintroduce empty elements to the solutions.

2) The conflicting element may be legally present in the both if()
branches, i.e. the conflicting element should be marked as deleted
regardless of if() condition.