lib: internal: nativeresolver: impl: __propose_solution: exclude rejection-only packages
When resolver selects a variant for a problem where more than 1 variant
is available, it may apply some rejects for it by (possibly) creating a new
entry with empty version and a list of elements which are no more
allowed to be placed into this solution.

It's not a mistake to report them to user, but there is no point doing
so.