debian: rules: override_dh_auto_install-arch: rebuild 'non-locally' first
This is done so that when debhelper proceeds to the testing stage, the
built cupt binaries/libs are built with LOCAL=on, so loading the
downloadmethod shared objects succeeds from the local path.

When we are done with testing, rebuild with LOCAL=off (which will
rebuild only one source file) and proceed with installation.